Découper en plusieurs morceaux le fichier de test de mes utilisateurs
