+++ src/components/MesUtilisateurs/MesUtilisateurs.test.tsx
@@ -1,13 +1,9 @@
 import { fireEvent, screen, waitFor, within } from '@testing-library/react'
-import { clearFirst, select } from 'react-select-event'
 import { Mock } from 'vitest'
 
 import MesUtilisateurs from './MesUtilisateurs'
-import departements from '../../../ressources/departements.json'
-import groupements from '../../../ressources/groupements.json'
-import regions from '../../../ressources/regions.json'
-import { renderComponent, matchWithoutMarkup, structuresFetch } from '@/components/testHelper'
-import { mesUtilisateursPresenter, RolesAvecStructure } from '@/presenters/mesUtilisateursPresenter'
+import { renderComponent, rolesAvecStructure } from '@/components/testHelper'
+import { mesUtilisateursPresenter } from '@/presenters/mesUtilisateursPresenter'
 import { sessionUtilisateurViewModelFactory } from '@/presenters/testHelper'
 import { utilisateurReadModelFactory } from '@/use-cases/testHelper'
 
@@ -421,848 +417,6 @@ describe('mes utilisateurs', () => {
     const navigation = screen.queryByRole('navigation', { name: 'Pagination' })
     expect(navigation).not.toBeInTheDocument()
   })
-
-  it('quand je clique sur le bouton pour filtrer alors les filtres apparaissent', () => {
-    // GIVEN
-    const mesUtilisateursViewModel = mesUtilisateursPresenter([utilisateurActifReadModel, utilisateurEnAttenteReadModel], 'fooId', totalUtilisateur, rolesAvecStructure)
-    renderComponent(<MesUtilisateurs mesUtilisateursViewModel={mesUtilisateursViewModel} />)
-
-    // WHEN
-    const filtrer = screen.getByRole('button', { name: 'Filtrer' })
-    fireEvent.click(filtrer)
-
-    // THEN
-    const drawerFiltrer = screen.getByRole('dialog', { name: 'Filtrer' })
-
-    const titre = within(drawerFiltrer).getByRole('heading', { level: 1, name: 'Filtrer' })
-    expect(titre).toBeInTheDocument()
-
-    const formulaire = within(drawerFiltrer).getByRole('form', { name: 'Filtrer' })
-    expect(formulaire).toHaveAttribute('method', 'dialog')
-
-    const zonesGeographiques = within(drawerFiltrer).getByRole('combobox', { name: 'Par zone géographique' })
-    expect(zonesGeographiques).toBeInTheDocument()
-
-    const administrateurDispositif = within(formulaire).getByLabelText('Administrateur dispositif')
-    expect(administrateurDispositif).toBeChecked()
-    const gestionnaireDepartement = within(formulaire).getByLabelText('Gestionnaire département')
-    expect(gestionnaireDepartement).toBeChecked()
-    const gestionnaireGroupement = within(formulaire).getByLabelText('Gestionnaire groupement')
-    expect(gestionnaireGroupement).toBeChecked()
-    const gestionnaireRegion = within(formulaire).getByLabelText('Gestionnaire région')
-    expect(gestionnaireRegion).toBeChecked()
-    const gestionnaireStructure = within(formulaire).getByLabelText('Gestionnaire structure')
-    expect(gestionnaireStructure).toBeChecked()
-    const instructeur = within(formulaire).getByLabelText('Instructeur')
-    expect(instructeur).toBeChecked()
-    const pilotePolitiquePublique = within(formulaire).getByLabelText('Pilote politique publique')
-    expect(pilotePolitiquePublique).toBeChecked()
-    const supportAnimation = within(formulaire).getByLabelText('Support animation')
-    expect(supportAnimation).toBeChecked()
-
-    const boutonReinitialiser = within(formulaire).getByRole('button', { name: 'Réinitialiser les filtres' })
-    expect(boutonReinitialiser).toHaveAttribute('type', 'reset')
-
-    const boutonAfficher = within(formulaire).getByRole('button', { name: 'Afficher les utilisateurs' })
-    expect(boutonAfficher).toBeEnabled()
-    expect(boutonAfficher).toHaveAttribute('type', 'submit')
-    expect(boutonAfficher).toHaveAttribute('aria-controls', 'drawer-filtre-utilisateurs')
-  })
-
-  it('ayant des filtres déjà actifs quand je clique sur le bouton pour filtrer alors ils apparaissent préremplis', () => {
-    // GIVEN
-    const mesUtilisateursViewModel = mesUtilisateursPresenter([utilisateurActifReadModel, utilisateurEnAttenteReadModel], 'fooId', totalUtilisateur, rolesAvecStructure)
-    renderComponent(
-      <MesUtilisateurs mesUtilisateursViewModel={mesUtilisateursViewModel} />,
-      { searchParams: new URLSearchParams('utilisateursActives=on&roles=gestionnaire_groupement,instructeur') }
-    )
-
-    // WHEN
-    const filtrer = screen.getByRole('button', { name: 'Filtrer' })
-    fireEvent.click(filtrer)
-    const filtres = screen.getByRole('dialog', { name: 'Filtrer' })
-
-    // THEN
-    const utilisateursActives = within(filtres).getByLabelText('Uniquement les utilisateurs activés')
-    expect(utilisateursActives).toBeChecked()
-    const administrateurDispositif = within(filtres).getByLabelText('Administrateur dispositif')
-    expect(administrateurDispositif).not.toBeChecked()
-    const gestionnaireDepartement = within(filtres).getByLabelText('Gestionnaire département')
-    expect(gestionnaireDepartement).not.toBeChecked()
-    const gestionnaireGroupement = within(filtres).getByLabelText('Gestionnaire groupement')
-    expect(gestionnaireGroupement).toBeChecked()
-    const gestionnaireRegion = within(filtres).getByLabelText('Gestionnaire région')
-    expect(gestionnaireRegion).not.toBeChecked()
-    const gestionnaireStructure = within(filtres).getByLabelText('Gestionnaire structure')
-    expect(gestionnaireStructure).not.toBeChecked()
-    const instructeur = within(filtres).getByLabelText('Instructeur')
-    expect(instructeur).toBeChecked()
-    const pilotePolitiquePublique = within(filtres).getByLabelText('Pilote politique publique')
-    expect(pilotePolitiquePublique).not.toBeChecked()
-    const supportAnimation = within(filtres).getByLabelText('Support animation')
-    expect(supportAnimation).not.toBeChecked()
-  })
-
-  it('quand je clique sur le bouton pour réinitialiser les filtres alors je repars de zéro', () => {
-    // GIVEN
-    const spiedRouterPush = vi.fn()
-    afficherLesFiltres(spiedRouterPush)
-
-    // WHEN
-    const boutonReinitialiser = screen.getByRole('button', { name: 'Réinitialiser les filtres' })
-    fireEvent.click(boutonReinitialiser)
-
-    // THEN
-    expect(spiedRouterPush).toHaveBeenCalledWith('/mes-utilisateurs')
-  })
-
-  describe('quand je filtre', () => {
-    it('sur les utilisateurs activés alors je n’affiche qu’eux', () => {
-      // GIVEN
-      const spiedRouterPush = vi.fn()
-      afficherLesFiltres(spiedRouterPush)
-      const filtres = screen.getByRole('dialog', { name: 'Filtrer' })
-      const utilisateursActives = within(filtres).getByLabelText('Uniquement les utilisateurs activés')
-      fireEvent.click(utilisateursActives)
-
-      // WHEN
-      const boutonAfficher = within(filtres).getByRole('button', { name: 'Afficher les utilisateurs' })
-      fireEvent.click(boutonAfficher)
-
-      // THEN
-      expect(spiedRouterPush).toHaveBeenCalledWith('http://example.com/mes-utilisateurs?utilisateursActives=on')
-    })
-
-    it('sur certains rôles alors je n’affiche qu’eux', () => {
-      // GIVEN
-      const spiedRouterPush = vi.fn()
-      afficherLesFiltres(spiedRouterPush)
-      const filtres = screen.getByRole('dialog', { name: 'Filtrer' })
-      const gestionnaireRegion = within(filtres).getByLabelText('Gestionnaire région')
-      fireEvent.click(gestionnaireRegion)
-      const gestionnaireDepartement = within(filtres).getByLabelText('Gestionnaire département')
-      fireEvent.click(gestionnaireDepartement)
-
-      // WHEN
-      const boutonAfficher = within(filtres).getByRole('button', { name: 'Afficher les utilisateurs' })
-      fireEvent.click(boutonAfficher)
-
-      // THEN
-      expect(spiedRouterPush).toHaveBeenCalledWith('http://example.com/mes-utilisateurs?roles=administrateur_dispositif%2Cgestionnaire_groupement%2Cgestionnaire_structure%2Cinstructeur%2Cpilote_politique_publique%2Csupport_animation')
-    })
-
-    it('sur un département alors je n’affiche qu’eux', async () => {
-      // GIVEN
-      const spiedRouterPush = vi.fn()
-      afficherLesFiltres(spiedRouterPush)
-      const zoneGeographique = screen.getByLabelText('Par zone géographique')
-      await select(zoneGeographique, '(978) Saint-Martin')
-
-      // WHEN
-      const boutonAfficher = screen.getByRole('button', { name: 'Afficher les utilisateurs' })
-      fireEvent.click(boutonAfficher)
-
-      // THEN
-      expect(spiedRouterPush).toHaveBeenCalledWith('http://example.com/mes-utilisateurs?codeDepartement=978')
-    })
-
-    it('sur une région alors je n’affiche qu’eux', async () => {
-      // GIVEN
-      const spiedRouterPush = vi.fn()
-      afficherLesFiltres(spiedRouterPush)
-      const zoneGeographique = screen.getByLabelText('Par zone géographique')
-      await select(zoneGeographique, "(93) Provence-Alpes-Côte d'Azur")
-
-      // WHEN
-      const boutonAfficher = screen.getByRole('button', { name: 'Afficher les utilisateurs' })
-      fireEvent.click(boutonAfficher)
-
-      // THEN
-      expect(spiedRouterPush).toHaveBeenCalledWith('http://example.com/mes-utilisateurs?codeRegion=93')
-    })
-
-    it('sur une structure alors je n’affiche que les utilisateurs liés à cette structure', async () => {
-      // GIVEN
-      vi.stubGlobal('fetch', structuresFetch)
-      const spiedRouterPush = vi.fn()
-      afficherLesFiltres(spiedRouterPush)
-      const filtreParStructure = screen.getByLabelText('Par structure')
-      fireEvent.input(filtreParStructure, { target: { value: 'tet' } })
-      await select(filtreParStructure, 'TETRIS')
-
-      // WHEN
-      const boutonAfficher = screen.getByRole('button', { name: 'Afficher les utilisateurs' })
-      fireEvent.click(boutonAfficher)
-
-      // THEN
-      expect(filtreParStructure).toBeInTheDocument()
-      expect(spiedRouterPush).toHaveBeenCalledWith('http://example.com/mes-utilisateurs?structure=14')
-    })
-
-    describe('sur une zone géographique et une structure', () => {
-      it.each([
-        {
-          desc: 'sur une région et une structure de cette région',
-          expectedRouterPush: 'http://example.com/mes-utilisateurs?codeRegion=93&structure=14',
-          zoneGeographique: "(93) Provence-Alpes-Côte d'Azur",
-        },
-        {
-          desc: 'sur un département et une structure de ce département',
-          expectedRouterPush: 'http://example.com/mes-utilisateurs?codeDepartement=06&structure=14',
-          zoneGeographique: '(06) Alpes-Maritimes',
-        },
-        {
-          desc: 'sur toutes les zones géographiques et une structure',
-          expectedRouterPush: 'http://example.com/mes-utilisateurs?structure=14',
-          zoneGeographique: 'Toutes les régions',
-        },
-
-      ])(
-        '$desc, alors je n’affiche que les utilisateurs liés à cette structure',
-        async ({ expectedRouterPush, zoneGeographique }) => {
-          // GIVEN
-          vi.stubGlobal('fetch', structuresFetch)
-          const spiedRouterPush = vi.fn()
-          afficherLesFiltres(spiedRouterPush)
-          await select(screen.getByLabelText('Par zone géographique'), zoneGeographique)
-          const filtreParStructure = screen.getByLabelText('Par structure')
-          fireEvent.input(filtreParStructure, { target: { value: 'tet' } })
-          await select(filtreParStructure, 'TETRIS')
-
-          // WHEN
-          const boutonAfficher = screen.getByRole('button', { name: 'Afficher les utilisateurs' })
-          fireEvent.click(boutonAfficher)
-
-          // THEN
-          expect(filtreParStructure).toBeInTheDocument()
-          expect(spiedRouterPush).toHaveBeenCalledWith(expectedRouterPush)
-        }
-      )
-
-      it('après avoir effacé la zone géographique précédemment sélectionnée, alors je n’affiche que les utilisateurs liés à cette structure', async () => {
-        // GIVEN
-        vi.stubGlobal('fetch', structuresFetch)
-        const spiedRouterPush = vi.fn()
-        afficherLesFiltres(spiedRouterPush)
-        const filtreParZoneGeographique = screen.getByLabelText('Par zone géographique')
-        await select(filtreParZoneGeographique, '(06) Alpes-Maritimes')
-        const filtreParStructure = screen.getByLabelText('Par structure')
-        fireEvent.input(filtreParStructure, { target: { value: 'tet' } })
-        await select(filtreParStructure, 'TETRIS')
-
-        // WHEN
-        const boutonAfficher = screen.getByRole('button', { name: 'Afficher les utilisateurs' })
-        await clearFirst(filtreParZoneGeographique)
-        fireEvent.click(boutonAfficher)
-
-        // THEN
-        expect(filtreParStructure).toBeInTheDocument()
-        expect(spiedRouterPush).toHaveBeenCalledWith('http://example.com/mes-utilisateurs?structure=14')
-      })
-    })
-  })
-
-  describe('quand j’invite un utilisateur', () => {
-    it('en tant qu’administrateur, quand je clique sur le bouton inviter, alors le drawer s’ouvre avec tous les rôles sélectionnables', async () => {
-      // GIVEN
-      const mesUtilisateursViewModel = mesUtilisateursPresenter([utilisateurActifReadModel, utilisateurEnAttenteReadModel], 'fooId', totalUtilisateur, rolesAvecStructure)
-      renderComponent(<MesUtilisateurs mesUtilisateursViewModel={mesUtilisateursViewModel} />, {
-        sessionUtilisateurViewModel: sessionUtilisateurViewModelFactory({
-          role: {
-            groupe: 'admin',
-            libelle: 'Rhône',
-            nom: 'Administrateur dispositif',
-            pictogramme: 'maille',
-            rolesGerables: [
-              'Administrateur dispositif',
-              'Gestionnaire département',
-              'Gestionnaire groupement',
-              'Gestionnaire région',
-              'Gestionnaire structure',
-              'Instructeur',
-              'Pilote politique publique',
-              'Support animation',
-            ],
-          },
-        }),
-      })
-      const inviter = screen.getByRole('button', { name: 'Inviter une personne' })
-
-      // WHEN
-      fireEvent.click(inviter)
-
-      // THEN
-      const drawerInvitation = screen.getByRole('dialog', { name: 'Invitez un utilisateur à rejoindre l’espace de gestion' })
-      const titre = await within(drawerInvitation).findByRole('heading', { level: 1, name: 'Invitez un utilisateur à rejoindre l’espace de gestion' })
-      expect(titre).toBeInTheDocument()
-
-      const champsObligatoires = within(drawerInvitation).getByText(
-        matchWithoutMarkup('Les champs avec * sont obligatoires.'),
-        { selector: 'p' }
-      )
-      expect(champsObligatoires).toBeInTheDocument()
-
-      const formulaireInvitation = screen.getByRole('form', { name: 'Inviter un utilisateur' })
-      expect(formulaireInvitation).toHaveAttribute('method', 'dialog')
-      const nom = within(formulaireInvitation).getByLabelText('Nom *')
-      expect(nom).toBeRequired()
-      expect(nom).toHaveAttribute('name', 'nom')
-      expect(nom).toHaveAttribute('type', 'text')
-
-      const prenom = within(formulaireInvitation).getByLabelText('Prénom *')
-      expect(prenom).toBeRequired()
-      expect(prenom).toHaveAttribute('name', 'prenom')
-      expect(prenom).toHaveAttribute('type', 'text')
-
-      const email = within(formulaireInvitation).getByLabelText('Adresse électronique *Une invitation lui sera envoyée par e-mail')
-      expect(email).toBeRequired()
-      expect(email).toHaveAttribute('name', 'email')
-      expect(email).toHaveAttribute('pattern', '.+@.+\\..{2,}')
-      expect(email).toHaveAttribute('type', 'email')
-      expect(email).toHaveAttribute('aria-describedby', 'text-input-error-desc-error')
-
-      const fieldset = within(formulaireInvitation).getByRole('group')
-      const roleQuestion = within(fieldset).getByText(
-        matchWithoutMarkup('Quel rôle souhaitez-vous lui attribuer ? *'),
-        { selector: 'legend' }
-      )
-      expect(roleQuestion).toBeInTheDocument()
-
-      const administrateurDispositif = within(fieldset).getByLabelText('Administrateur dispositif')
-      expect(administrateurDispositif).toBeRequired()
-      expect(administrateurDispositif).toHaveAttribute('name', 'attributionRole')
-      expect(administrateurDispositif).toHaveAttribute('id', 'Administrateur dispositif')
-
-      const gestionnaireRegion = within(fieldset).getByLabelText('Gestionnaire région')
-      expect(gestionnaireRegion).toBeRequired()
-      expect(gestionnaireRegion).toHaveAttribute('name', 'attributionRole')
-      expect(gestionnaireRegion).toHaveAttribute('id', 'Gestionnaire région')
-
-      const gestionnaireDepartement = within(fieldset).getByLabelText('Gestionnaire département')
-      expect(gestionnaireDepartement).toBeRequired()
-      expect(gestionnaireDepartement).toHaveAttribute('name', 'attributionRole')
-      expect(gestionnaireDepartement).toHaveAttribute('id', 'Gestionnaire département')
-
-      const gestionnaireGroupement = within(fieldset).getByLabelText('Gestionnaire groupement')
-      expect(gestionnaireGroupement).toBeRequired()
-      expect(gestionnaireGroupement).toHaveAttribute('name', 'attributionRole')
-      expect(gestionnaireGroupement).toHaveAttribute('id', 'Gestionnaire groupement')
-
-      const gestionnaireStructure = within(fieldset).getByLabelText('Gestionnaire structure')
-      expect(gestionnaireStructure).toBeRequired()
-      expect(gestionnaireStructure).toHaveAttribute('name', 'attributionRole')
-      expect(gestionnaireStructure).toHaveAttribute('id', 'Gestionnaire structure')
-
-      const instructeur = within(fieldset).getByLabelText('Instructeur')
-      expect(instructeur).toBeRequired()
-      expect(instructeur).toHaveAttribute('name', 'attributionRole')
-      expect(instructeur).toHaveAttribute('id', 'Instructeur')
-
-      const pilotePolitiquePublique = within(fieldset).getByLabelText('Pilote politique publique')
-      expect(pilotePolitiquePublique).toBeRequired()
-      expect(pilotePolitiquePublique).toHaveAttribute('name', 'attributionRole')
-      expect(pilotePolitiquePublique).toHaveAttribute('id', 'Pilote politique publique')
-
-      const supportAnimation = within(fieldset).getByLabelText('Support animation')
-      expect(supportAnimation).toBeRequired()
-      expect(supportAnimation).toHaveAttribute('name', 'attributionRole')
-      expect(supportAnimation).toHaveAttribute('id', 'Support animation')
-
-      const envoyerInvitation = within(formulaireInvitation).getByRole('button', { name: 'Envoyer l’invitation' })
-      expect(envoyerInvitation).toHaveAttribute('type', 'submit')
-    })
-
-    it('en tant qu’administrateur, quand je clique sur un rôle à inviter, alors le champ d’organisation s’affiche', () => {
-      // GIVEN
-      const mesUtilisateursViewModel = mesUtilisateursPresenter([utilisateurActifReadModel, utilisateurEnAttenteReadModel], 'fooId', totalUtilisateur, rolesAvecStructure)
-      renderComponent(<MesUtilisateurs mesUtilisateursViewModel={mesUtilisateursViewModel} />, {
-        sessionUtilisateurViewModel: sessionUtilisateurViewModelFactory({
-          role: {
-            groupe: 'admin',
-            libelle: 'Rhône',
-            nom: 'Administrateur dispositif',
-            pictogramme: 'maille',
-            rolesGerables: [
-              'Administrateur dispositif',
-              'Gestionnaire département',
-              'Gestionnaire groupement',
-              'Gestionnaire région',
-              'Gestionnaire structure',
-              'Instructeur',
-              'Pilote politique publique',
-              'Support animation',
-            ],
-          },
-        }),
-      })
-      const inviter = screen.getByRole('button', { name: 'Inviter une personne' })
-      fireEvent.click(inviter)
-
-      // WHEN
-      const formulaireInvitation = screen.getByRole('dialog', { name: 'Invitez un utilisateur à rejoindre l’espace de gestion' })
-      const gestionnaireDepartement = within(formulaireInvitation).getByLabelText('Gestionnaire département')
-      fireEvent.click(gestionnaireDepartement)
-
-      // THEN
-      const organisation = within(formulaireInvitation).getByLabelText('Département *')
-      expect(organisation).toBeRequired()
-      expect(organisation).toHaveAttribute('type', 'text')
-    })
-
-    it('en tant qu’administrateur, quand je fais une recherche dans le champ de structure, alors je peux y faire une recherche utilisée dans le formulaire', async () => {
-      // GIVEN
-      const inviterUnUtilisateurAction = vi.fn(async () => Promise.resolve(['OK']))
-      const windowDsfr = window.dsfr
-      window.dsfr = (): {modal: {conceal: Mock}} => ({
-        modal: {
-          conceal: vi.fn(),
-        },
-      })
-      vi.stubGlobal('fetch', vi.fn(structuresFetch))
-      const mesUtilisateursViewModel = mesUtilisateursPresenter([utilisateurActifReadModel, utilisateurEnAttenteReadModel], 'fooId', totalUtilisateur, rolesAvecStructure)
-      renderComponent(<MesUtilisateurs mesUtilisateursViewModel={mesUtilisateursViewModel} />, {
-        inviterUnUtilisateurAction,
-        sessionUtilisateurViewModel: sessionUtilisateurViewModelFactory({
-          role: {
-            groupe: 'admin',
-            libelle: 'Rhône',
-            nom: 'Administrateur dispositif',
-            pictogramme: 'maille',
-            rolesGerables: [
-              'Administrateur dispositif',
-              'Gestionnaire département',
-              'Gestionnaire groupement',
-              'Gestionnaire région',
-              'Gestionnaire structure',
-              'Instructeur',
-              'Pilote politique publique',
-              'Support animation',
-            ],
-          },
-        }),
-      })
-      const inviter = screen.getByRole('button', { name: 'Inviter une personne' })
-      fireEvent.click(inviter)
-      const formulaireInvitation = screen.getByRole('dialog', { name: 'Invitez un utilisateur à rejoindre l’espace de gestion' })
-      const nom = within(formulaireInvitation).getByLabelText('Nom *')
-      fireEvent.change(nom, { target: { value: 'Tartempion' } })
-      const prenom = within(formulaireInvitation).getByLabelText('Prénom *')
-      fireEvent.change(prenom, { target: { value: 'Martin' } })
-      const email = within(formulaireInvitation).getByLabelText(/Adresse électronique/)
-      fireEvent.change(email, { target: { value: 'martin.tartempion@example.com' } })
-      const gestionnaireStructure = within(formulaireInvitation).getByLabelText('Gestionnaire structure')
-      fireEvent.click(gestionnaireStructure)
-
-      // WHEN
-      const structure = within(formulaireInvitation).getByLabelText('Structure *')
-      fireEvent.change(structure, { target: { value: 'ABC' } })
-      await select(structure, 'ABC FORMATION')
-      const envoyerInvitation = await within(formulaireInvitation).findByRole('button', { name: 'Envoyer l’invitation' })
-      fireEvent.click(envoyerInvitation)
-
-      // THEN
-      await waitFor(() => {
-        expect(inviterUnUtilisateurAction).toHaveBeenCalledWith({
-          codeOrganisation: '1845',
-          email: 'martin.tartempion@example.com',
-          nom: 'Tartempion',
-          prenom: 'Martin',
-          role: 'Gestionnaire structure',
-        })
-      })
-      expect(fetch).toHaveBeenCalledWith('/api/structures?search=ABC')
-      window.dsfr = windowDsfr
-    })
-
-    it('en tant qu’administrateur, quand je fais une recherche de moins de 3 caractères dans le champ de structure, alors il ne se passe rien', () => {
-      // GIVEN
-      vi.stubGlobal('fetch', vi.fn(() => ({ json: async (): Promise<ReadonlyArray<{
-        nom: string
-        uid: string
-      }>> => Promise.resolve([]) })))
-      const mesUtilisateursViewModel = mesUtilisateursPresenter([utilisateurActifReadModel, utilisateurEnAttenteReadModel], 'fooId', totalUtilisateur, rolesAvecStructure)
-      renderComponent(<MesUtilisateurs mesUtilisateursViewModel={mesUtilisateursViewModel} />, {
-        sessionUtilisateurViewModel: sessionUtilisateurViewModelFactory({
-          role: {
-            groupe: 'admin',
-            libelle: 'Rhône',
-            nom: 'Administrateur dispositif',
-            pictogramme: 'maille',
-            rolesGerables: [
-              'Administrateur dispositif',
-              'Gestionnaire département',
-              'Gestionnaire groupement',
-              'Gestionnaire région',
-              'Gestionnaire structure',
-              'Instructeur',
-              'Pilote politique publique',
-              'Support animation',
-            ],
-          },
-        }),
-      })
-
-      // WHEN
-      const inviter = screen.getByRole('button', { name: 'Inviter une personne' })
-      fireEvent.click(inviter)
-      const formulaireInvitation = screen.getByRole('dialog', { name: 'Invitez un utilisateur à rejoindre l’espace de gestion' })
-      const gestionnaireStructure = within(formulaireInvitation).getByLabelText('Gestionnaire structure')
-      fireEvent.click(gestionnaireStructure)
-      const structure = screen.getByLabelText('Structure *')
-      fireEvent.change(structure, { target: { value: 'AB' } })
-
-      // THEN
-      expect(fetch).not.toHaveBeenCalled()
-    })
-
-    it('en tant que gestionnaire département, quand je clique sur le bouton inviter, alors le drawer s’ouvre avec tous le rôle gestionnaire département sélectionné', async () => {
-      // GIVEN
-      const mesUtilisateursViewModel = mesUtilisateursPresenter([utilisateurActifReadModel, utilisateurEnAttenteReadModel], 'fooId', totalUtilisateur, rolesAvecStructure)
-      renderComponent(<MesUtilisateurs mesUtilisateursViewModel={mesUtilisateursViewModel} />, {
-        sessionUtilisateurViewModel: sessionUtilisateurViewModelFactory({
-          role: {
-            groupe: 'gestionnaire',
-            libelle: 'Rhône',
-            nom: 'Gestionnaire département',
-            pictogramme: 'maille',
-            rolesGerables: ['Gestionnaire département'],
-          },
-        }),
-      })
-      const inviter = screen.getByRole('button', { name: 'Inviter une personne' })
-
-      // WHEN
-      fireEvent.click(inviter)
-
-      // THEN
-      const formulaireInvitation = screen.getByRole('dialog', { name: 'Invitez un utilisateur à rejoindre l’espace de gestion' })
-      const titre = await within(formulaireInvitation).findByRole('heading', { level: 1, name: 'Invitez un utilisateur à rejoindre l’espace de gestion' })
-      expect(titre).toBeInTheDocument()
-
-      const champsObligatoires = within(formulaireInvitation).getByText(
-        matchWithoutMarkup('Les champs avec * sont obligatoires.'),
-        { selector: 'p' }
-      )
-      expect(champsObligatoires).toBeInTheDocument()
-
-      const nom = within(formulaireInvitation).getByLabelText('Nom *')
-      expect(nom).toBeRequired()
-      expect(nom).toHaveAttribute('name', 'nom')
-      expect(nom).toHaveAttribute('type', 'text')
-
-      const prenom = within(formulaireInvitation).getByLabelText('Prénom *')
-      expect(prenom).toBeRequired()
-      expect(prenom).toHaveAttribute('name', 'prenom')
-      expect(prenom).toHaveAttribute('type', 'text')
-
-      const email = within(formulaireInvitation).getByLabelText('Adresse électronique *Une invitation lui sera envoyée par e-mail')
-      expect(email).toBeRequired()
-      expect(email).toHaveAttribute('name', 'email')
-      expect(email).toHaveAttribute('pattern', '.+@.+\\..{2,}')
-      expect(email).toHaveAttribute('type', 'email')
-      expect(email).toHaveAttribute('aria-describedby', 'text-input-error-desc-error')
-
-      const roleQuestion = within(formulaireInvitation).getByText(
-        matchWithoutMarkup('Rôle attribué à cet utilisateur :'),
-        { selector: 'p' }
-      )
-      expect(roleQuestion).toBeInTheDocument()
-
-      const role = within(formulaireInvitation).getByText(
-        matchWithoutMarkup('Gestionnaire département'),
-        { selector: 'p' }
-      )
-      expect(role).toBeInTheDocument()
-
-      const envoyerInvitation = within(formulaireInvitation).getByRole('button', { name: 'Envoyer l’invitation' })
-      expect(envoyerInvitation).toHaveAttribute('type', 'submit')
-    })
-
-    it('quand je remplis correctement le formulaire et avec un nouveau mail, alors un message de validation s’affiche et le drawer est réinitialisé', async () => {
-      // GIVEN
-      const inviterUnUtilisateurAction = vi.fn(async () => Promise.resolve(['OK']))
-      const windowDsfr = window.dsfr
-      window.dsfr = (): {modal: {conceal: Mock}} => ({
-        modal: {
-          conceal: vi.fn(),
-        },
-      })
-      const mesUtilisateursViewModel = mesUtilisateursPresenter([utilisateurActifReadModel, utilisateurEnAttenteReadModel], 'fooId', totalUtilisateur, rolesAvecStructure)
-      renderComponent(
-        <MesUtilisateurs mesUtilisateursViewModel={mesUtilisateursViewModel} />, {
-          inviterUnUtilisateurAction,
-          sessionUtilisateurViewModel: sessionUtilisateurViewModelFactory({
-            role: {
-              groupe: 'admin',
-              libelle: 'Rhône',
-              nom: 'Administrateur dispositif',
-              pictogramme: 'maille',
-              rolesGerables: [
-                'Administrateur dispositif',
-                'Gestionnaire département',
-                'Gestionnaire groupement',
-                'Gestionnaire région',
-                'Gestionnaire structure',
-                'Instructeur',
-                'Pilote politique publique',
-                'Support animation',
-              ],
-            },
-          }),
-        }
-      )
-      const inviter = screen.getByRole('button', { name: 'Inviter une personne' })
-      fireEvent.click(inviter)
-      const formulaireInvitation = screen.getByRole('dialog', { name: 'Invitez un utilisateur à rejoindre l’espace de gestion' })
-      const roleRadios = within(formulaireInvitation).getAllByRole('radio')
-
-      // WHEN
-      const nom = within(formulaireInvitation).getByLabelText('Nom *')
-      fireEvent.change(nom, { target: { value: 'Tartempion' } })
-      const prenom = within(formulaireInvitation).getByLabelText('Prénom *')
-      fireEvent.change(prenom, { target: { value: 'Martin' } })
-      const email = within(formulaireInvitation).getByLabelText(/Adresse électronique/)
-      fireEvent.change(email, { target: { value: 'martin.tartempion@example.com' } })
-      const administrateurDispositif = within(formulaireInvitation).getByLabelText('Administrateur dispositif')
-      fireEvent.click(administrateurDispositif)
-      const envoyerInvitation = await within(formulaireInvitation).findByRole('button', { name: 'Envoyer l’invitation' })
-      fireEvent.click(envoyerInvitation)
-
-      // THEN
-      const notification = await screen.findByRole('alert')
-      expect(notification).toHaveTextContent('Invitation envoyée à martin.tartempion@example.com')
-      expect(formulaireInvitation).not.toHaveAttribute('open', '')
-      expect(nom).toHaveValue('')
-      expect(prenom).toHaveValue('')
-      expect(email).toHaveValue('')
-      roleRadios.forEach((roleRadio) => {
-        expect(roleRadio).not.toBeChecked()
-      })
-      window.dsfr = windowDsfr
-    })
-
-    it('quand je remplis avec un e-mail déjà existant puis avec un e-mail inexistant le formulaire d’invitation, alors un message d’invalidation puis de validation s’affiche et le drawer est réinitialisé', async () => {
-      // GIVEN
-      const roleGestionnaireLabelSelectionMapping = {
-        'Gestionnaire département': 'Département',
-        'Gestionnaire groupement': 'Groupement',
-        'Gestionnaire région': 'Région',
-        'Gestionnaire structure': 'Structure',
-      }
-      const inviterUnUtilisateurAction = vi.fn()
-        .mockResolvedValueOnce(['emailExistant'])
-        .mockResolvedValueOnce(['OK'])
-      const windowDsfr = window.dsfr
-      window.dsfr = (): {modal: {conceal: Mock}} => ({
-        modal: {
-          conceal: vi.fn(),
-        },
-      })
-      const mesUtilisateursViewModel = mesUtilisateursPresenter([utilisateurActifReadModel, utilisateurEnAttenteReadModel], 'fooId', totalUtilisateur, rolesAvecStructure)
-      renderComponent(
-        <MesUtilisateurs mesUtilisateursViewModel={mesUtilisateursViewModel} />, {
-          inviterUnUtilisateurAction,
-          sessionUtilisateurViewModel: sessionUtilisateurViewModelFactory({
-            role: {
-              groupe: 'admin',
-              libelle: 'Rhône',
-              nom: 'Administrateur dispositif',
-              pictogramme: 'maille',
-              rolesGerables: [
-                'Administrateur dispositif',
-                'Gestionnaire département',
-                'Gestionnaire groupement',
-                'Gestionnaire région',
-                'Gestionnaire structure',
-                'Instructeur',
-                'Pilote politique publique',
-                'Support animation',
-              ],
-            },
-          }),
-        }
-      )
-      const inviter = screen.getByRole('button', { name: 'Inviter une personne' })
-      fireEvent.click(inviter)
-      const formulaireInvitation = screen.getByRole('dialog', { name: 'Invitez un utilisateur à rejoindre l’espace de gestion' })
-      const roleRadios = within(formulaireInvitation).getAllByRole('radio')
-
-      // WHEN
-      const nom = within(formulaireInvitation).getByLabelText('Nom *')
-      fireEvent.change(nom, { target: { value: 'Tartempion' } })
-      const prenom = within(formulaireInvitation).getByLabelText('Prénom *')
-      fireEvent.change(prenom, { target: { value: 'Martin' } })
-      const email = within(formulaireInvitation).getByLabelText(/Adresse électronique/)
-      fireEvent.change(email, { target: { value: 'martin.tartempion@example.com' } })
-      const gestionnaireRole = within(formulaireInvitation).getByLabelText('Gestionnaire département')
-      fireEvent.click(gestionnaireRole)
-      const departementSelect = within(formulaireInvitation).getByLabelText('Département *')
-      expect(departementSelect).toBeInTheDocument()
-      await select(departementSelect, 'Ain')
-      const envoyerInvitation = await within(formulaireInvitation).findByRole('button', { name: 'Envoyer l’invitation' })
-      fireEvent.click(envoyerInvitation)
-      const messageDErreur = await within(formulaireInvitation).findByText('Cet utilisateur dispose déjà d’un compte', { selector: 'p' })
-      fireEvent.click(envoyerInvitation)
-
-      // THEN
-      expect(messageDErreur).toBeInTheDocument()
-      const absenceMessageDErreur = await within(formulaireInvitation).findByText('Cet utilisateur dispose déjà d’un compte', { selector: 'p' })
-      expect(absenceMessageDErreur).not.toBeInTheDocument()
-      const notification = await screen.findByRole('alert')
-      expect(notification).toHaveTextContent('Invitation envoyée à martin.tartempion@example.com')
-      expect(formulaireInvitation).not.toHaveAttribute('open', '')
-      expect(nom).toHaveValue('')
-      expect(prenom).toHaveValue('')
-      expect(email).toHaveValue('')
-      roleRadios.forEach((roleRadio) => {
-        expect(roleRadio).not.toBeChecked()
-      })
-      Object.values(roleGestionnaireLabelSelectionMapping).forEach((labelChampSelection) => {
-        const champSelection = within(formulaireInvitation).queryByLabelText(`${labelChampSelection} *`)
-        expect(champSelection).not.toBeInTheDocument()
-      })
-      window.dsfr = windowDsfr
-    })
-
-    it('dans le drawer d’invitation, quand je remplis correctement le formulaire et avec un mail existant, alors il y a un message d’erreur', async () => {
-      // GIVEN
-      const inviterUnUtilisateurAction = vi.fn(async () => Promise.resolve(['emailExistant']))
-      const mesUtilisateursViewModel = mesUtilisateursPresenter([utilisateurActifReadModel, utilisateurEnAttenteReadModel], 'fooId', totalUtilisateur, rolesAvecStructure)
-      renderComponent(
-        <MesUtilisateurs mesUtilisateursViewModel={mesUtilisateursViewModel} />,
-        {
-          inviterUnUtilisateurAction,
-          sessionUtilisateurViewModel: sessionUtilisateurViewModelFactory({
-            role: {
-              groupe: 'admin',
-              libelle: 'Rhône',
-              nom: 'Administrateur dispositif',
-              pictogramme: 'maille',
-              rolesGerables: [
-                'Administrateur dispositif',
-                'Gestionnaire département',
-                'Gestionnaire groupement',
-                'Gestionnaire région',
-                'Gestionnaire structure',
-                'Instructeur',
-                'Pilote politique publique',
-                'Support animation',
-              ],
-            },
-          }),
-        }
-      )
-      const inviter = screen.getByRole('button', { name: 'Inviter une personne' })
-      fireEvent.click(inviter)
-      const formulaireInvitation = screen.getByRole('dialog', { name: 'Invitez un utilisateur à rejoindre l’espace de gestion' })
-
-      // WHEN
-      const nom = within(formulaireInvitation).getByLabelText('Nom *')
-      fireEvent.change(nom, { target: { value: 'Tartempion' } })
-      const prenom = within(formulaireInvitation).getByLabelText('Prénom *')
-      fireEvent.change(prenom, { target: { value: 'Martin' } })
-      const email = within(formulaireInvitation).getByLabelText(/Adresse électronique/)
-      fireEvent.change(email, { target: { value: 'martin.tartempion@example.com' } })
-      const administrateurDispositif = within(formulaireInvitation).getByLabelText('Administrateur dispositif')
-      fireEvent.click(administrateurDispositif)
-      const envoyerInvitation = await within(formulaireInvitation).findByRole('button', { name: 'Envoyer l’invitation' })
-      fireEvent.click(envoyerInvitation)
-
-      // THEN
-      const erreurEmailDejaExistant = await within(formulaireInvitation).findByText('Cet utilisateur dispose déjà d’un compte', { selector: 'p' })
-      expect(erreurEmailDejaExistant).toBeInTheDocument()
-      expect(formulaireInvitation).toHaveAttribute('open', '')
-    })
-  })
-
-  it('dans le drawer d’invitation, quand je remplis correctement le formulaire mais que l’invitation ne peut pas se faire, alors le drawer se ferme', async () => {
-    // GIVEN
-    const inviterUnUtilisateurAction = vi.fn(async () => Promise.resolve(['KO']))
-    const windowDsfr = window.dsfr
-    // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-    window.dsfr = () => ({
-      modal: {
-        conceal: vi.fn(),
-      },
-    })
-    const mesUtilisateursViewModel = mesUtilisateursPresenter([utilisateurActifReadModel], 'fooId', totalUtilisateur, rolesAvecStructure)
-    renderComponent(
-      <MesUtilisateurs mesUtilisateursViewModel={mesUtilisateursViewModel} />,
-      {
-        inviterUnUtilisateurAction,
-        sessionUtilisateurViewModel: sessionUtilisateurViewModelFactory({
-          role: {
-            groupe: 'admin',
-            libelle: 'Rhône',
-            nom: 'Administrateur dispositif',
-            pictogramme: 'maille',
-            rolesGerables: [
-              'Administrateur dispositif',
-              'Gestionnaire département',
-              'Gestionnaire groupement',
-              'Gestionnaire région',
-              'Gestionnaire structure',
-              'Instructeur',
-              'Pilote politique publique',
-              'Support animation',
-            ],
-          },
-        }),
-      }
-    )
-    const inviter = screen.getByRole('button', { name: 'Inviter une personne' })
-    fireEvent.click(inviter)
-    const formulaireInvitation = screen.getByRole('dialog', { name: 'Invitez un utilisateur à rejoindre l’espace de gestion' })
-    const roleRadios = within(formulaireInvitation).getAllByRole('radio')
-
-    // WHEN
-    const nom = within(formulaireInvitation).getByLabelText('Nom *')
-    fireEvent.change(nom, { target: { value: 'Tartempion' } })
-    const prenom = within(formulaireInvitation).getByLabelText('Prénom *')
-    fireEvent.change(prenom, { target: { value: 'Martin' } })
-    const email = within(formulaireInvitation).getByLabelText(/Adresse électronique/)
-    fireEvent.change(email, { target: { value: 'martin.tartempion@example.com' } })
-    const administrateurDispositif = within(formulaireInvitation).getByLabelText('Administrateur dispositif')
-    fireEvent.click(administrateurDispositif)
-    const envoyerInvitation = await within(formulaireInvitation).findByRole('button', { name: 'Envoyer l’invitation' })
-    fireEvent.click(envoyerInvitation)
-
-    // THEN
-    const absenceDeMessageDErreur = within(formulaireInvitation).queryByText('Cet utilisateur dispose déjà d’un compte', { selector: 'p' })
-    expect(absenceDeMessageDErreur).not.toBeInTheDocument()
-    const formulaireInvitationApresInvitationEnEchec = await screen.findByRole(
-      'dialog',
-      { name: 'Invitez un utilisateur à rejoindre l’espace de gestion' }
-    )
-    expect(formulaireInvitationApresInvitationEnEchec).not.toHaveAttribute('open', '')
-    expect(nom).toHaveValue('')
-    expect(prenom).toHaveValue('')
-    expect(email).toHaveValue('')
-    roleRadios.forEach((roleRadio) => {
-      expect(roleRadio).not.toBeChecked()
-    })
-    window.dsfr = windowDsfr
-  })
-
-  function afficherLesFiltres(spiedRouterPush: Mock): void {
-    const mesUtilisateursViewModel = mesUtilisateursPresenter([utilisateurActifReadModel, utilisateurEnAttenteReadModel], 'fooId', totalUtilisateur, rolesAvecStructure)
-    renderComponent(
-      <MesUtilisateurs mesUtilisateursViewModel={mesUtilisateursViewModel} />,
-      {
-        router: {
-          back: vi.fn(),
-          forward: vi.fn(),
-          prefetch: vi.fn(),
-          push: spiedRouterPush,
-          refresh: vi.fn(),
-          replace: vi.fn(),
-        },
-      }
-    )
-
-    const filtrer = screen.getByRole('button', { name: 'Filtrer' })
-    fireEvent.click(filtrer)
-  }
 })
 
 function getByTable(): { columnsHead: ReadonlyArray<HTMLElement>, rowsBody: ReadonlyArray<HTMLElement> } {
@@ -1333,22 +487,3 @@ const utilisateurActifSansTelephoneVideReadModel = utilisateurReadModelFactory({
   },
   telephone: '',
 })
-
-const rolesAvecStructure: RolesAvecStructure = {
-  'Gestionnaire département': {
-    label: 'Département',
-    options: departements.map((departement) => ({ label: departement.nom, value: departement.code })),
-  },
-  'Gestionnaire groupement': {
-    label: 'Groupement',
-    options: groupements.map((groupement) => ({ label: groupement.nom, value: `${groupement.id}` })),
-  },
-  'Gestionnaire région': {
-    label: 'Région',
-    options: regions.map((region) => ({ label: region.nom, value: region.code })),
-  },
-  'Gestionnaire structure': {
-    label: 'Structure',
-    options: [],
-  },
-}
